rest/server/pyproject.toml
[tool.poetry]
name = "diva_server"
version = "0.4.0"
description = "Server logic for DiVA REST service"
authors = ["Shin Saito <shinsa@jp.ibm.com>"]
license = "EPL"

[tool.poetry.dependencies]
python = "3.7.*"
PyYAML = "^5.4.1"
connexion = { extras = ["swagger-ui"], version = "^2.9.0" }
aiohttp = "^3.7.4"
aiohttp-jinja2 = "^1.4.2"
pytest = "^6.2.4"
requests = "^2.26.0"
importlib-metadata = "^4.6.4"
Flask-Cors = "^3.0.10"
docker = "^5.0.0"
environs = "^9.3.3"

[tool.poetry.dev-dependencies]
autopep8 = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
